+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## 1.0.6 - 2026-03-30
 ### Fixed
 - Fixed INDI shared driver
+- Fixed issues with compatibility layer
 
 ## 1.0.5 - 2026-03-27
 ### Fixed